+++ models/Page.js
@@ -13,7 +13,12 @@ class Page extends Sequelize.Model {
                     allowNull: false,
                     charset: 'utf8mb4',
                     collate: 'utf8mb4_unicode_ci'
-                }
+                },
+                deleted: {
+                    type: Sequelize.BOOLEAN,
+                    allowNull: false,
+                    defaultValue: false,
+                },
             }, 
             {   // 테이블 설정
                 sequelize,
@@ -21,9 +26,6 @@ class Page extends Sequelize.Model {
                 underscored: true, // 테이블명과 컬럼명을 camelCase, snake_case 선택
                 modelName: 'Page', // 모델이름
                 tableName: 'pages', // 테이블이름
-                paranoid: true, // true이면 deletedAt컬럼이 자동으로 생성되고 삭제시 삭제하지 않음
-                // charset: 'utf8', // 인코딩
-                // collate: 'utf8_general_ci', // 정렬시 비교기준
                 charset: 'utf8mb4',
                 collate: 'utf8mb4_unicode_ci'
             }
